Refactor search functionality to useSearch
